+++ src/test/java/org/sonar/plugins/delphi/pmd/rules/PublicFieldsRuleTest.java
@@ -44,4 +44,23 @@ public class PublicFieldsRuleTest extends BasePmdRuleTest {
         .hasSize(1)
         .areExactly(1, ruleKeyAtLine("PublicFieldsRule", builder.getOffsetDecl() + 9));
   }
+
+  @Test
+  public void testRecordsAreExcluded() {
+    DelphiTestUnitBuilder builder = new DelphiTestUnitBuilder();
+    builder.appendDecl("type");
+    builder.appendDecl("  TMyRecord = record");
+    builder.appendDecl("     FPublishedField: Integer;");
+    builder.appendDecl("    private");
+    builder.appendDecl("     FPrivateField: Integer;");
+    builder.appendDecl("    protected");
+    builder.appendDecl("     FProtectedField: String;");
+    builder.appendDecl("    public");
+    builder.appendDecl("     FPublicField: String;");
+    builder.appendDecl("  end;");
+
+    execute(builder);
+
+    assertIssues().isEmpty();
+  }
 }