+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Exclude records in `PublicFieldsRule`.
+
 ## [0.16.2] - 2020-04-22
 
 ### Changed